+++ import_export/admin.py
@@ -329,7 +329,7 @@ class ImportMixin(BaseImportMixin, ImportExportMixinBase):
             warnings.warn(
                 "ImportMixin.get_confirm_import_form() is deprecated and will "
                 "be removed in a future release. Please use the new "
-                "`confirm_form_class` attribute to specify a custom form "
+                "'confirm_form_class' attribute to specify a custom form "
                 "class, or override the get_confirm_form_class() method if "
                 "your requirements are more complex.",
                 category=DeprecationWarning
@@ -365,7 +365,7 @@ class ImportMixin(BaseImportMixin, ImportExportMixinBase):
             'import_file_name': import_form.cleaned_data["import_file"].tmp_storage_name,
             'original_file_name': import_form.cleaned_data["import_file"].name,
             'input_format': import_form.cleaned_data["input_format"],
-            'resource':import_form.cleaned_data.get("resource", ""),
+            'resource': import_form.cleaned_data.get("resource", ""),
         }
 
     def get_import_data_kwargs(self, request, *args, **kwargs):
@@ -600,22 +600,17 @@ class ExportMixin(BaseExportMixin, ImportExportMixinBase):
             Use :meth:`~import_export.admin.ExportMixin.get_export_form_class` or set the new
             :attr:`~import_export.admin.ExportMixin.export_form_class` attribute.
         """
+        warnings.warn(
+            "ExportMixin.get_export_form() is deprecated and will be removed in "
+            "a future release. Please use get_export_form_class() instead.",
+            category=DeprecationWarning
+        )
         return self.export_form_class
 
     def get_export_form_class(self):
         """
         Get the form class used to read the export format.
         """
-        if not getattr(self.get_export_form, 'is_original', False):
-            warnings.warn(
-                "ExportMixin.get_export_form_class() is deprecated and will "
-                "be removed in a future release. Please use the new "
-                "`export_form_class` attribute to specify a custom form "
-                "class, or override the get_export_form_class() method if "
-                "your requirements are more complex.",
-                category=DeprecationWarning
-            )
-            return self.get_confirm_import_form()
         return self.export_form_class
 
     def export_action(self, request, *args, **kwargs):